[SPARK-50838][SQL] Performs additional checks inside recursive CTEs to throw an error if forbidden case is encountered
### What changes were proposed in this pull request?

Performs additional checks inside recursive CTEs to throw an error if forbidden case is encountered:
1. Recursive term can contain one recursive reference only.
2. Recursive reference can't be used in some kinds of joins and aggregations.
3. Recursive references are not allowed in subqueries

In addition, the name of `recursive` function inside `CTERelationDef` is rewritten to `hasRecursiveCTERelationRef` and adds `hasItsOwnUnionLoopRef` function as it is also needed to check if cteDef is recursive after substitution.

A small bug in `CTESubstitution` is fixed which now enables substitution of self-references within subqueries as well (but not its resolution, as they are not allowed).

### Why are the changes needed?

Support for the recursive CTE.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

will be tested in https://github.com/apache/spark/pull/49571

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49518 from milanisvet/checkRecursion.

Authored-by: Milan Cupac <milan.cupac@outlook.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit ea60a59e7da675fc9dbd97a46676bef29ea0b67d)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>