[SPARK-49700][CONNECT][SQL] Unified Scala Interface for Connect and Classic
### What changes were proposed in this pull request?
This PR makes the shared SQL (JVM) interface the primary interface for Scala/JVM based Dataframe programming.

The implementations are moved to the `classic` and `connect` sub packages.

The connect client had to be moved to the sql/connect/common package because serialization requires the captured client classes to be on the classpath when deserialized on the server.

### Why are the changes needed?
This is the final step in creating a unified Scala interface for both Classic and Connect.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48818 from hvanhovell/SPARK-49700.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>
(cherry picked from commit 09e94a1f66a5f0924f28f6f94468604a1f1d332f)
Signed-off-by: Herman van Hovell <herman@databricks.com>