[MINOR][SQL] Remove signature from Hive thriftserver DEBUG log
### What changes were proposed in this pull request?

This PR aims to remove signature from Hive thriftserver DEBUG log.

### Why are the changes needed?

This is aligned with the upstream Apache Hive project.
- https://github.com/apache/hive/pull/4887

This is also aligned with the previous Spark change.
- #43402

### Does this PR introduce _any_ user-facing change?

No, this is a debug message removal.

### How was this patch tested?

Pass the CIs and manual review (because this is a removal).

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49752 from dongjoon-hyun/minor_debug_msg.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 04e0ddf96cedb0ebc8f0d4439562f90df174e8ae)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>