[SPARK-51011][CORE] Add logging for whether a task is going to be interrupted when killed
We now log the value of `interruptThread` when a `TaskRunner`'s `kill` method is killed. This should help with debugging when potential zombie Spark tasks do not seem to be exiting.

Today, it's tricky to debug why a task is not exiting (and thus, why executors might be getting lost) without knowing for sure if it was issued a Java interrupt.

No.

Ran `org.apache.spark.executor.ExecutorSuite` and verified the log looked as expected.

No.

Closes #49699 from neilramaswamy/spark-51011.

Lead-authored-by: Neil Ramaswamy <neil.ramaswamy@databricks.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit fbf5e7bdcb9775efeba878882de4a0b657323958)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>