[SPARK-51011][CORE] Add logging for whether a task is going to be interrupted when killed
### What changes were proposed in this pull request?

We now log the value of `interruptThread` when a `TaskRunner`'s `kill` method is killed. This should help with debugging when potential zombie Spark tasks do not seem to be exiting.

### Why are the changes needed?

Today, it's tricky to debug why a task is not exiting (and thus, why executors might be getting lost) without knowing for sure if it was issued a Java interrupt.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Ran `org.apache.spark.executor.ExecutorSuite` and verified the log looked as expected.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49699 from neilramaswamy/spark-51011.

Lead-authored-by: Neil Ramaswamy <neil.ramaswamy@databricks.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>