[SPARK-51034][SQL] Reformat Describe As JSON statistics dict for parse-ability
### What changes were proposed in this pull request?

Reformat `Describe As JSON` statistics table metadata into a dict rather than string for improved parse-ability

### Why are the changes needed?

Existing `Describe` formats table statistics into a string `xxx bytes, xxx rows` which is not conducive to parsing. This PR stores the table statistics metadata in raw format and delegates the formatting to the caller of toJsonLinkedHashmap, for improved parsing of the JSON statistics.

### Does this PR introduce _any_ user-facing change?

Yes, it affects the output of `Describe As JSON`.

### How was this patch tested?

Added to test suite `DescribeTableSuite`

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49728 from asl3/asl3/describetable-statistics.

Authored-by: Amanda Liu <amanda.liu@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>