[SPARK-49308][CONNECT] Support UserDefinedAggregateFunction in Spark Connect Scala Client
### What changes were proposed in this pull request?
This PR adds support for `UserDefinedAggregateFunction` to the Spark Connect Scala Client. While this is a deprecated feature, we still believe it is useful to support it to ensure we reduce incompatibilities between classic and connect.

Implementation wise I opted to convert the `UserDefinedAggregateFunction` to an `Aggregator`, and use that code path for execution. This is probably not as fast as the original implementation (more allocations).

### Why are the changes needed?
This reduces friction between the classic and connect implementations.

### Does this PR introduce _any_ user-facing change?
Yes. It enabled Spark Connect Scala Client users to use `UserDefinedAggregateFunction`s.

### How was this patch tested?
Added tests to

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #49785 from hvanhovell/SPARK-49308.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>
(cherry picked from commit f69a07370e49ab56a5fdd593a6aaeae8d904dafd)
Signed-off-by: Herman van Hovell <herman@databricks.com>