[SPARK-51084][SQL] Assign appropriate error class for `negativeScaleNotAllowedError`
### What changes were proposed in this pull request?

Clarify error message for `negativeScaleNotAllowedError` with user-facing error class

### Why are the changes needed?

Raise a more user-friendly error message when users attempt to use a negative precision for DecimalType when the Spark config `spark.sql.legacy.allowNegativeScaleOfDecimal` is not set. Previously an internal error was raised, which is not correct.

### Does this PR introduce _any_ user-facing change?

Yes, it affects the error class when users attempt to use a negative precision for DecimalType when the Spark config `spark.sql.legacy.allowNegativeScaleOfDecimal` is not set.

### How was this patch tested?

Added test in `DecimalSuite.scala`

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49807 from asl3/asl3/negativescalenotallowederror.

Authored-by: Amanda Liu <amanda.liu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>