[MINOR][DOCS] Fix incorrect description of constraint on spark.sql.adaptive.coalescePartitions.minPartitionSize
### What changes were proposed in this pull request?

This PR addresses a minor problem in the SQL performance guide's description of the `spark.sql.adaptive.coalescePartitions.minPartitionSize` configuration.

Currently, the guide says:

> The minimum size of shuffle partitions after coalescing. Its value can be at most 20% of `spark.sql.adaptive.advisoryPartitionSizeInBytes`. This is useful when the target size is ignored during partition coalescing, which is the default case.

but the second sentence is not true because the 20% limitation was removed in https://github.com/apache/spark/pull/33655#discussion_r683582184 (SPARK-36430), which shipped in 3.2.0. That PR updated the in-code description but not the guide on the web.

### Why are the changes needed?

Incorrect doc is misleading.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

n/a

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49872 from JoshRosen/fix-spark.sql.adaptive.coalescePartitions.minPartitionSize-doc.

Authored-by: Josh Rosen <joshrosen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 06a11f3d003fcba2ea33a5f03dc9cf660e368ee6)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>