[SPARK-51154][BUILD][TESTS] Remove unused `jopt` test dependency
### What changes were proposed in this pull request?

The Kafka connector declares a dependecy on the jopt library. This library is not actually used in the code, so we can remove it.

### Why are the changes needed?

We have no apparent need for this dependency. The version we are including is 15 years old with numerous unresolved bugs.

### Does this PR introduce _any_ user-facing change?

No. This change only impacts tests.

### How was this patch tested?

Run all Kafka tests.

```
build/mvn -pl connector/kafka-0-10,connector/kafka-0-10-sql test
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49877 from cnauroth/SPARK-51154.

Authored-by: Chris Nauroth <cnauroth@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>