[SPARK-51185][CORE] Revert simplifications to PartitionedFileUtil API to reduce memory requirements
### What changes were proposed in this pull request?

This PR reverts an earlier change (https://github.com/apache/spark/pull/41632) that converted FileStatusWithMetadata.getPath from a def to a lazy val in order to simplify the PartitionedFileUtils helpers.

### Why are the changes needed?

The conversion of getPath from a def to a lazy val increases the memory requirements because now paths need to be kept in memory as long as the FileStatusWithMetadata exists. As paths are expensive to store, this can lead to higher memory utilization and increase the risk for OOMs.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

This is a small revert to code that has already existed before so the existing tests are sufficient.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #49915 from LukasRupprecht/def_get-path.

Authored-by: Lukas Rupprecht <lukas.l.rupprecht@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit eac7b305b35986190c64a4f0bee04a605843cfde)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>