[SPARK-51198][CORE][DOCS] Revise `defaultMinPartitions` function description
### What changes were proposed in this pull request?

This PR aims to revise the function description of `SparkContext.defaultMinPartitions`.

### Why are the changes needed?

The repository, `https://github.com/mesos/spark`, was archived 11 years ago.

We had better embed the description instead of pointing to the ancient repository. It's hard to figure out the real reasons from a series of comments.

https://github.com/apache/spark/blob/356174076748c434b6ef35df4b58a9ecdf04f727/core/src/main/scala/org/apache/spark/SparkContext.scala#L2880

### Does this PR introduce _any_ user-facing change?

No. This is a change of comment.

### How was this patch tested?

Manual review.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49929 from dongjoon-hyun/SPARK-51198.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>