[SPARK-51183][SQL] Link to Parquet spec in Variant docs
### What changes were proposed in this pull request?

The Parquet spec in https://github.com/apache/parquet-format/blob/master/VariantEncoding.md is based on the one in Spark, but has received a number of updates (especially related to Shredding).

At this point, the code in Spark more closely matches the latest version of the Parquet spec (the main gap being the lack of a few new scalar types that were recently added, and which we will try to add to Spark soon).

This PR updates the README.md and shredding.md files to just point to the Parquet spec, which we plan to have the Spark code follow.

### Why are the changes needed?

Improve internal documentation and avoid maintaining two copies of the spec.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

None.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49910 from cashmand/fix_variant_readme.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>