[SPARK-51083][CORE] Modify JavaUtils to not swallow InterruptedExceptions
### What changes were proposed in this pull request?

These changes modify the `deleteRecursivelyUsingUnixNative` method in `JavaUtils.java` to not swallow `InterruptedException`s. The bulk of the changes in this PR relate to modifying the signatures of methods that directly or indirectly use `JavaUtils.deleteRecursivelyUsingUnixNative`.

### Why are the changes needed?

`JavaUtils.deleteRecursively` swallows `InterruptedException`s, and that can cause entire executor loss. Consider a streaming task running a streaming aggregation. It takes the following steps:

1. It [writes](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/statefulOperators.scala#L720) all data to the state store, and then [reads](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/statefulOperators.scala#L732) this data back to emit it (in, say, Complete mode).
2. With the RocksDB state store, it doesn't actually acquire any locks, perform any sleeping, etc. during (1). This means that if the query is cancelled and it is interrupted, it won't respond to the interrupt.
3. After (1) is done, it [calls commit](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/statefulOperators.scala#L744) on the state store.
4. In RocksDB, this turns into a call to `RocksDB::commit`, which can call [createSnapshot](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/state/RocksDB.scala#L879).
5. In createSnapshot, [we call Utils.deleteRecursively](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/state/RocksDB.scala#L879) to make sure that the temporary directory into which we place the RocksDB checkpoint is empty.
6. In deleteRecursively, we [call deleteRecursivelyUsingUnixNative](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/common/utils/src/main/java/org/apache/spark/network/util/JavaUtils.java#L115) which issues an `rm -rf` and then [calls process.waitFor](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/common/utils/src/main/java/org/apache/spark/network/util/JavaUtils.java#L183), which declares that it could throw an InterruptedException.
7. Unfortunately, deleteRecursivelyUsingUnixNative [indiscriminately catches all Exceptions](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/common/utils/src/main/java/org/apache/spark/network/util/JavaUtils.java#L184), which includes the `InterruptedException` potentially thrown by `process.waitFor`. Then, it rethrows an IOException.
8. That IOException is caught [here](https://github.com/apache/spark/blob/048484116857eb3d1156d139ab439e990ae078c6/common/utils/src/main/java/org/apache/spark/network/util/JavaUtils.java#L117), and a warning is logged. The interrupt has now been swallowed.

A streaming task thread that misses this interrupted exception will now not exit. If it doesn't exit for 60 seconds, then the TaskReaper will kill the JVM.

### Does this PR introduce _any_ user-facing change?

Yes.

### How was this patch tested?

This is fairly tricky to test; in fact, we don't exercise the code-path for `deleteRecursivelyUsingUnixNative` during testing on Apple Silicon, which made it hard for me.

My approach here was to modify `deleteRecursivelyUsingUnixNative` to explicitly sleep. Then, I called that method in a background thread, interrupting it from the test thread. If all went correctly, then the background thread _should_ have received an interrupt that it did not swallow, and the test thread could assert that that happened. You can see this test in [this commit](https://github.com/apache/spark/pull/49796/commits/95577db37c9209e64034db2e22879590ae636020), but since it's an intrusive change, I have reverted it. Example output is:

```
[info] UtilsSuite:
1. Starting thread
2. Waiting for it to get to its sleep
3. Going to spawn rm -rf
4. Finished deleting directory
5. Sleeping for 10 seconds
6. Interrupting thread that is sleeping
[info] - deleteRecursively throws InterruptedException (5 seconds, 18 milliseconds)
7. Catching and rethrowing interrupted exception
8. Thread exiting
9. gotInterruptedException = true
```

I validated that this test case I added _fails_ without my changes to fix the interrupt swallowing; you can check out that commit [here](https://github.com/apache/spark/pull/49796/commits/70f2a7b3e916a7a54b23901919159adc49ab8dda).

If you are a reviewer who can think of a way to create a test case we can check-in, please leave a review comment.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #49796 from neilramaswamy/nr/spark-51083.

Authored-by: Neil Ramaswamy <neil.ramaswamy@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>