[SPARK-51294][SQL][CONNECT][TESTS] Improve the readability by split the variable of jars and configs for SparkConnectServerUtils
### What changes were proposed in this pull request?
This PR proposes to improve the readability by split the variable of jars and configs for `SparkConnectServerUtils`.

### Why are the changes needed?
Currently, according to the name and the description of variable `testConfigs`, it is for tests that needs some special configs. But it contains a parameter about test jar.
Reference the variable `connectJar`, we should extract the parameter of test jar to a single variable `catalystTestJar`.

### Does this PR introduce _any_ user-facing change?
'No'.

### How was this patch tested?
GA.

### Was this patch authored or co-authored using generative AI tooling?
'No'.

Closes #50050 from beliefer/SPARK-51294.

Authored-by: beliefer <beliefer@163.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 11f6edd33f6dd308882845ed1419f991d41a734e)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>