[SPARK-49912] Refactor simple CASE statement to evaluate the case variable only once
### What changes were proposed in this pull request?
In this PR, CASE statement is refactored. Existing `CaseStatement` is split into two - `SimpleCaseStatement` and `SearchedCaseStatement`. `SearchedCaseStatement` retains the old behavior, while for `SimpleCaseStatement` a new logical and execution node are added - `SimpleCaseStatement` and `SimpleCaseStatementExec`.

Previously, a simple case statement would evaluate the case variable again for every WHEN clause in the CASE. This is both inefficient, and could produce unexpected behavior if the evaluation has a side effect. `SimpleCaseStatementExec` now caches the result of the evaluation, and compares the WHEN conditions to the cached `Literal`.

### Why are the changes needed?
Previous iteration of simple CASE evaluated the expression multiple times.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Tests were added to `SqlScriptingParserSuite`, `SqlScriptingInterpreterSuite` and `SqlScriptingExecutionNodeSuite`.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50027 from dusantism-db/scripting-case-improvements-v2.

Authored-by: Dušan Tišma <dusan.tisma@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>