[SPARK-51315][SQL] Enabling object level collations by default
### What changes were proposed in this pull request?

Enabling the flag for object level collations by default, as the feature is now working end-to-end.
Also, a minor fix is made in DDL command resolution for AlterTableCommand, ensuring that the altered object is an instance of ResolvedTable, to avoid exposing this path for other types such as ResolvedPersistentView. This is just additional safety mechanism.

### Why are the changes needed?

The feature is now working end-to-end, and can be enabled accordingly.

### Does this PR introduce _any_ user-facing change?

No. Previous PRs related to this functionality already introduced all underlying user facing changes.

### How was this patch tested?

Existing dedicated tests already cover this functionality (the flag was already enabled in testing environment).

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50082 from dejankrak-db/enable-object-level-collations.

Lead-authored-by: Dejan Krakovic <dejan.krakovic@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>