[SPARK-51302][CONNECT] Spark Connect supports JDBC should use the DataFrameReader API
### What changes were proposed in this pull request?
This PR proposes to unify the calling to the `DataFrameReader` API in Spark Connect where supports the jdbc API.

### Why are the changes needed?
The origin code is good at a little advance of performance, but it is bad if we change the logic of jdbc API.
I think we should unify the code path here.

### Does this PR introduce _any_ user-facing change?
'No'.

### How was this patch tested?
GA.

### Was this patch authored or co-authored using generative AI tooling?
'No'.

Closes #50059 from beliefer/SPARK-51302.

Authored-by: beliefer <beliefer@163.com>
Signed-off-by: beliefer <beliefer@163.com>